// frontend/src/app/models/FinishedJobs.ts
export class FinishedJobs {
    _id: String = ""
    endDate: Date = new Date(1, 0, 1);
    username: String = "";
    firm: String = "";
    pictures: String[] = [];
    owner: String = "";
    comment: String = "";
    rating: Number = -1;
    poolNumber: number = 0;
    commentOwner: String = "";
    fountainNumber: number = 0;
    lastMaintained: Date = new Date();
}